requirements.txt
bottom==2.1.3
pyyaml==3.12
sqlalchemy==1.2.5
click==6.7
requests==2.18.4
websockets==4.0.1
tornado==5.0.1
aiohttp==3.0.0b0
cchardet==2.1.1
aiodns==1.1.1
sqlalchemy_aio==0.12.0
pymysql==0.8.1
discord==0.16.12
python-dateutil==2.7.3